virtual-serverに対応したテストを追加しました。
