+++ tests/unit_tests/HTTPRequest_test/HTTPRequest_test.cpp
@@ -1,5 +1,6 @@
 #include <gtest/gtest.h>
 #include <iostream>
+#include <arpa/inet.h>
 #include "./HTTPRequest.hpp"
 #include "./ClientSocket.hpp"
 #include "ListenSocket.hpp"
@@ -7,297 +8,142 @@
 #include "EventQueue.hpp"
 #include "HTTPServer.hpp"
 
-/*
 class RequestTest : public ::testing::Test
 {
-  protected:
-	virtual void SetUp()
-	{
-		Config	config("./default.conf");
-
-		const std::vector<ServerDirective>&				servers = config.GetServers();
-		std::vector<ServerDirective>::const_iterator	itr = servers.begin();
-
-		lsocket_ = new ListenSocket(*itr);
-		lsocket_->ListenConnection();
-		csocket_ = new ClientSocket();
-		csocket_->ConnectServer("127.0.0.1", 8080);
-		ssocket_ = new ServerSocket(lsocket_->AcceptConnection(), lsocket_->GetServerConf());
-	}
-
-	virtual void TearDown()
-	{
-		delete lsocket_;
-		delete ssocket_;
-		delete csocket_;
-	}
-
-	ListenSocket *lsocket_;
-	ServerSocket *ssocket_;
-	ClientSocket *csocket_;
-};
-*/
-
-class RequestTest : public ::testing::Test
-{
-  protected:
-	virtual void SetUp()
-	{
-		config_ = new Config("./default.conf");
-		equeue_ = new EventQueue();
-
-		RegisterListenSockets(*config_, equeue_);
-		MainLoop(*equeue_);
-	}
-
-	virtual void TearDown()
-	{
-		delete config_;
-		delete equeue_;
-	}
-
-
-
-	void	MainLoop(const EventQueue& equeue)
-	{
-		while (1)
+	protected:
+		static void SetUpTestCase()
 		{
-			udata = equeue.WaitEvent();
-			asocket = static_cast<ASocket*>(udata);
-			lsocket = dynamic_cast<ListenSocket*>(asocket);
-			ssocket = dynamic_cast<ServerSocket*>(asocket);
-			if (lsocket)
+			MyRegisterListenSockets(config_);
+		}
+    	static void TearDownTestCase()
+		{
+			std::vector<ListenSocket*>::const_iterator	itr = lsockets_.begin();
+			std::vector<ListenSocket*>::const_iterator	end = lsockets_.end();
+			while (itr != end)
 			{
-				std::cout << "Accept!!" << std::endl;
-				new_ssocket = new ServerSocket(*lsocket);
-				equeue.RegisterEvent(new_ssocket->GetFd(), new_ssocket);
+				delete *itr;
+				++itr;
 			}
-			else
-				Communication(ssocket);
 		}
-	}
-
-	void	Communication(const ServerSocket *ssocket)
-	{
-		int	status_code = 0;
-		req = new HTTPRequest(*ssocket);
-
-		try
+		virtual void TearDown()
 		{
-			req.ParseRequest();
-			req.RequestDisplay();
+			delete req_;
+			delete ssocket_;
 		}
-		catch (const ClientClosed& e)
+
+		static void	MyRegisterListenSockets(const Config& config)
 		{
-			delete ssocket;
-			return;
+			const std::vector<ServerDirective>&				servers = config.GetServers();
+			std::vector<ServerDirective>::const_iterator	sitr = servers.begin();
+			std::vector<ServerDirective>::const_iterator	send = servers.end();
+			std::vector<ListenSocket*>::const_iterator		same_listen_lsocket;
+			ListenSocket*									new_lsocket;
+
+			while (sitr != send)
+			{
+				const std::vector<ServerDirective::Listen>&				listens = sitr->GetListen();
+				std::vector<ServerDirective::Listen>::const_iterator	litr = listens.begin();
+				std::vector<ServerDirective::Listen>::const_iterator	lend = listens.end();
+				while (litr != lend)
+				{
+					same_listen_lsocket = Utils::FindMatchMember(lsockets_, &ListenSocket::GetListen, *litr);
+					if (same_listen_lsocket == lsockets_.end())
+					{
+						new_lsocket = new ListenSocket(*litr, *sitr);
+						lsockets_.push_back(new_lsocket);
+						new_lsocket->ListenConnection();
+					}
+					else
+						(*same_listen_lsocket)->AddServerConf(*sitr);
+					++litr;
+				}
+				++sitr;
+			}
 		}
-		catch (const HTTPError& e)
+
+		void	RunCommunication(const std::string& msg, const int port, const in_addr_t host = INADDR_ANY)
 		{
-			status_code = e.GetStatusCode();
+			ClientSocket	csocket;
+			csocket.ConnectServer("127.0.0.1", port);
+
+			std::vector<ListenSocket*>::const_iterator	target_lsocket;
+			const ServerDirective::Listen				listen = std::make_pair(host, port);
+			target_lsocket = Utils::FindMatchMember(lsockets_, &ListenSocket::GetListen, listen);
+
+			ssocket_ = new ServerSocket(**target_lsocket);
+			req_ = new HTTPRequest(*ssocket_);
+			try
+			{
+				csocket.SendRequest(msg);
+				req_->ParseRequest();
+				status_code_ = 200;
+			}
+			catch (const HTTPError& e)
+			{
+				status_code_ = e.GetStatusCode();
+			}
 		}
-		std::cout << "status_code: " << status_code << std::endl;
-	}
 
-	Config			*config_;
-	EventQueue		*equeue_;
+		static Config						config_;
+		static std::vector<ListenSocket*>	lsockets_;
 
-	void			*udata;
-	ASocket			*asocket;
-	ListenSocket	*lsocket;
-	ServerSocket	*ssocket;
-	ServerSocket 	*new_ssocket;
-
-	HTTPRequest		*req;
+		int									status_code_;
+		HTTPRequest*						req_;
+		ServerSocket*						ssocket_;
 };
 
-TEST_F(RequestTest, test)
+Config						RequestTest::config_("default.conf");
+std::vector<ListenSocket*>	RequestTest::lsockets_;
+
+TEST_F(RequestTest, test1)
 {
-	ClientSocket*	csocket_;
-
-	csocket_ = new ClientSocket();
-	csocket_->ConnectServer("127.0.0.1", 8080);
-
-	const char	*msg1 = "GET / HTTP/1.1\r\nHost: localhost:8080\r\n\r\n";
-	csocket_->SendRequest(msg1);
-
-	EXPECT_EQ("GET / HTTP/1.1", req.GetLine());
-	EXPECT_EQ("Host: localhost:8080", req.GetLine());
-	EXPECT_EQ("", req.GetLine());
-
-	const char	*msg2 = "aaaaa\r\nbbbbb\r\nccccc\r\n";
-	csocket_->SendRequest(msg2);
-
-	EXPECT_EQ("aaaaa", req.GetLine());
-	EXPECT_EQ("bbbbb", req.GetLine());
-	EXPECT_EQ("ccccc", req.GetLine());
+	RunCommunication("GET / HTTP/1.1\r\nHost: localhost:8080\r\n\r\n", 8080);
+	EXPECT_EQ("GET", req_->GetMethod());
+	EXPECT_EQ("/", req_->GetTarget());
+	EXPECT_EQ("HTTP/1.1", req_->GetVersion());
+	EXPECT_EQ("localhost", req_->GetHost().first);
+	EXPECT_EQ("8080", req_->GetHost().second);
+	EXPECT_EQ(INADDR_ANY, req_->GetListen().first);
+	EXPECT_EQ(8080, req_->GetListen().second);
+	EXPECT_EQ("webserv1", req_->GetServerConf()->GetServerNames()[0]);
+	EXPECT_EQ("default", req_->GetServerConf()->GetServerNames()[1]);
 }
 
-/*
-TEST_F(RequestTest, ParseMethodTest)
+TEST_F(RequestTest, test2)
 {
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseMethod("GET");
-	EXPECT_EQ("GET", req1.GetMethod());
-
-	HTTPRequest		req2(*ssocket_, ssocket_->GetServerConf());
-	req2.ParseMethod("POST");
-	EXPECT_EQ("POST", req2.GetMethod());
-
-	HTTPRequest		req3(*ssocket_, ssocket_->GetServerConf());
-	req3.ParseMethod("DELETE");
-	EXPECT_EQ("DELETE", req3.GetMethod());
-
-	HTTPRequest		req4(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req4.ParseMethod("test"));
-
-	HTTPRequest		req5(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req5.ParseMethod("TEST["));
+	RunCommunication("GET / HTTP/1.1\r\nHost: localhost:8081\r\n\r\n", 8081);
+	EXPECT_EQ("GET", req_->GetMethod());
+	EXPECT_EQ("/", req_->GetTarget());
+	EXPECT_EQ("HTTP/1.1", req_->GetVersion());
+	EXPECT_EQ("localhost", req_->GetHost().first);
+	EXPECT_EQ("8081", req_->GetHost().second);
+	EXPECT_EQ(INADDR_ANY, req_->GetListen().first);
+	EXPECT_EQ(8081, req_->GetListen().second);
+	EXPECT_EQ("webserv1", req_->GetServerConf()->GetServerNames()[0]);
+	EXPECT_EQ("default", req_->GetServerConf()->GetServerNames()[1]);
 }
 
-TEST_F(RequestTest, ParseTargetTest)
+TEST_F(RequestTest, test3)
 {
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseTarget("/");
-	EXPECT_EQ("/", req1.GetTarget());
-
-	HTTPRequest		req2(*ssocket_, ssocket_->GetServerConf());
-	req2.ParseTarget("/index.html");
-	EXPECT_EQ("/index.html", req2.GetTarget());
+	RunCommunication("GET / HTTP/1.1\r\nHost: localhost:8090\r\n\r\n", 8090);
+	EXPECT_EQ("GET", req_->GetMethod());
+	EXPECT_EQ("/", req_->GetTarget());
+	EXPECT_EQ("HTTP/1.1", req_->GetVersion());
+	EXPECT_EQ("localhost", req_->GetHost().first);
+	EXPECT_EQ("8090", req_->GetHost().second);
+	EXPECT_EQ(INADDR_ANY, req_->GetListen().first);
+	EXPECT_EQ(8090, req_->GetListen().second);
+	EXPECT_EQ("webserv3", req_->GetServerConf()->GetServerNames()[0]);
 }
-
-TEST_F(RequestTest, ParseVersionTest)
+TEST_F(RequestTest, test4)
 {
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseVersion("HTTP/1.1");
-	EXPECT_EQ("HTTP/1.1", req1.GetVersion());
-
-	HTTPRequest		req2(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req2.ParseVersion("HTTP/1.0"));
-
-	HTTPRequest		req3(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req3.ParseVersion("HTTP/0.0"));
-
-	HTTPRequest		req4(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req4.ParseVersion("HTTP/1.11"));
+	RunCommunication("GET / HTTP/1.1\r\nHost: webserv2:8080\r\n\r\n", 8080);
+	EXPECT_EQ("GET", req_->GetMethod());
+	EXPECT_EQ("/", req_->GetTarget());
+	EXPECT_EQ("HTTP/1.1", req_->GetVersion());
+	EXPECT_EQ("webserv2", req_->GetHost().first);
+	EXPECT_EQ("8080", req_->GetHost().second);
+	EXPECT_EQ(INADDR_ANY, req_->GetListen().first);
+	EXPECT_EQ(8080, req_->GetListen().second);
+	EXPECT_EQ("webserv2", req_->GetServerConf()->GetServerNames()[0]);
 }
-
-TEST_F(RequestTest, ParseHostTest)
-{
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseHost(" localhost:8080");
-	EXPECT_EQ("localhost:8080", req1.GetHost());
-
-	HTTPRequest		req2(*ssocket_, ssocket_->GetServerConf());
-	req2.ParseHost(" developer.mozilla.org");
-	EXPECT_EQ("developer.mozilla.org", req2.GetHost());
-}
-
-TEST_F(RequestTest, ParseContentLength)
-{
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseContentLength(" 123");
-	EXPECT_EQ((size_t)123, req1.GetContentLength());
-
-	HTTPRequest		req2(*ssocket_, ssocket_->GetServerConf());
-	req2.ParseContentLength(" 0");
-	EXPECT_EQ((size_t)0, req2.GetContentLength());
-
-	HTTPRequest		req3(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req3.ParseContentLength(" 12345abc"));
-
-	HTTPRequest		req4(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req4.ParseContentLength(""));
-
-	HTTPRequest		req5(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req5.ParseContentLength("         "));
-
-	HTTPRequest		req6(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req6.ParseContentLength("-1234"));
-
-	HTTPRequest		req7(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req7.ParseContentLength("123 456 789"));
-
-	HTTPRequest		req8(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req8.ParseContentLength("123,456,789"));
-
-	HTTPRequest		req10(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_ANY_THROW(req10.ParseContentLength("111111111111"));
-}
-
-TEST_F(RequestTest, ParseUserAgent)
-{
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseUserAgent("   Debian    ");
-	EXPECT_EQ("Debian", req1.GetUserAgent());
-}
-
-TEST_F(RequestTest, ParseAcceptEncodingTest)
-{
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	req1.ParseAcceptEncoding(" gzip, compress, br");
-	EXPECT_EQ("gzip", req1.GetAcceptEncoding().at(0));
-	EXPECT_EQ("compress", req1.GetAcceptEncoding().at(1));
-	EXPECT_EQ("br", req1.GetAcceptEncoding().at(2));
-}
-
-TEST_F(RequestTest, ParseConnectionTest)
-{
-	HTTPRequest		req1(*ssocket_, ssocket_->GetServerConf());
-	EXPECT_EQ(true, req1.GetConnection());
-	
-	HTTPRequest		req2(*ssocket_, ssocket_->GetServerConf());
-	req2.ParseConnection("close");
-	EXPECT_EQ(false, req2.GetConnection());
-
-	HTTPRequest		req3(*ssocket_, ssocket_->GetServerConf());
-	req3.ParseConnection("CLOSE");
-	EXPECT_EQ(false, req3.GetConnection());
-
-	HTTPRequest		req4(*ssocket_, ssocket_->GetServerConf());
-	req4.ParseConnection("Close");
-	EXPECT_EQ(false, req4.GetConnection());
-
-	HTTPRequest		req5(*ssocket_, ssocket_->GetServerConf());
-	req5.ParseConnection("cloSE");
-	EXPECT_EQ(false, req5.GetConnection());
-}
-
-TEST_F(RequestTest, GetLine)
-{
-	HTTPRequest		req(*ssocket_, ssocket_->GetServerConf());
-
-	const char	*msg1 = "GET / HTTP/1.1\r\nHost: localhost:8080\r\n\r\n";
-	csocket_->SendRequest(msg1);
-
-	EXPECT_EQ("GET / HTTP/1.1", req.GetLine());
-	EXPECT_EQ("Host: localhost:8080", req.GetLine());
-	EXPECT_EQ("", req.GetLine());
-
-	const char	*msg2 = "aaaaa\r\nbbbbb\r\nccccc\r\n";
-	csocket_->SendRequest(msg2);
-
-	EXPECT_EQ("aaaaa", req.GetLine());
-	EXPECT_EQ("bbbbb", req.GetLine());
-	EXPECT_EQ("ccccc", req.GetLine());
-}
-
-TEST_F(RequestTest, RequestTest)
-{
-	HTTPRequest		req(*ssocket_, ssocket_->GetServerConf());
-
-	const char	*msg1 = "POST / HTTP/1.1\r\nHost: localhost:8080\r\nUser-Agent: debian\r\nContent-Type: text/html\r\nContent-Length: 10\r\n\r\naaaaaaaaaa";
-	csocket_->SendRequest(msg1);
-
-	req.ParseRequest();
-	EXPECT_EQ("POST", req.GetMethod());
-	EXPECT_EQ("/", req.GetTarget());
-	EXPECT_EQ("HTTP/1.1", req.GetVersion());
-	EXPECT_EQ("localhost:8080", req.GetHost());
-	EXPECT_EQ((size_t)10, req.GetContentLength());
-	EXPECT_EQ(true, req.GetConnection());
-	EXPECT_EQ("debian", req.GetUserAgent());
-	EXPECT_EQ("text/html", req.GetContentType());
-	EXPECT_EQ("aaaaaaaaaa", req.GetBody());
-}
-*/
