+++ tests/unit_tests/HTTPRequest_test/Makefile
@@ -7,7 +7,7 @@ SRCS	=	HTTPRequest.cpp \
 			ListenSocket.cpp \
 			ServerSocket.cpp \
 			utils.cpp \
-			HTTPRequest_test_2.cpp \
+			HTTPRequest_test.cpp \
 			ClientSocket.cpp \
 			Config.cpp \
 			LocationDirective.cpp \
@@ -25,7 +25,9 @@ INCLUDE =	-I $(SRCSDIR)/server -I $(SRCSDIR)/socket -I $(SRCSDIR)/utils -I $(SRC
 NAME	=	unit_test
 
 CC		=	c++
+#CC		=	clang++
 CFLAGS	=	-Wall -Wextra -Werror
+#CFLAGS	=	-g -fsanitize=address
 CFLAGS	+=	-MMD -MP
 CFLAGS	+=	-std=c++11
 LINK	=	-lgtest -lgtest_main
