+++ tests/unit_tests/HTTPRequest_test/default.conf
@@ -44,6 +44,7 @@ server {
 
 server {
 	listen		8090;
+	server_name	webserv3;
 
 	location / {
 		return 301 http://localhost:8080;
